move JsonObjectUtils to its own class.
